Add trpc config and setup for /app
