+++ app/api/auth/[...nextauth]/route.ts
@@ -1,5 +1,6 @@
 import NextAuth from "next-auth";
-import authOptions from "../authOptions";
+
+import { authOptions } from "@/server/auth";
 
 const handler = NextAuth(authOptions);
