+++ app/api/trpc/[trpc]/route.ts
@@ -1,0 +1,26 @@
+import { fetchRequestHandler } from "@trpc/server/adapters/fetch";
+import { type NextRequest } from "next/server";
+import { appRouter } from "@/server/api/router";
+import { createTRPCContext } from "@/server/api/trpc";
+
+const NODE_ENV = process.env.NODE_ENV;
+
+if (!NODE_ENV) throw new Error("No NODE_ENV found.");
+
+const handler = (req: NextRequest) =>
+  fetchRequestHandler({
+    endpoint: "/api/trpc",
+    req,
+    router: appRouter,
+    createContext: () => createTRPCContext({ req }),
+    onError:
+      NODE_ENV === "development"
+        ? ({ path, error }) => {
+            console.error(
+              `❌ tRPC failed on ${path ?? "<no-path>"}: ${error.message}`,
+            );
+          }
+        : undefined,
+  });
+
+export { handler as GET, handler as POST };
