+++ components/Theme/ThemeProvider.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { ReactNode } from "react";
 import { ThemeProvider } from "next-themes";
