+++ server/api/router/auth.ts
@@ -1,0 +1,7 @@
+import { createTRPCRouter, publicProcedure } from "../trpc";
+
+export const authRouter = createTRPCRouter({
+  getSession: publicProcedure.query(({ ctx }) => {
+    return ctx.session;
+  }),
+});
