+++ server/api/router/index.ts
@@ -1,11 +1,11 @@
-import { router } from "../trpc";
+import { createTRPCRouter } from "../trpc";
 import { postRouter } from "./post";
 import { profileRouter } from "./profile";
 import { commentRouter } from "./comment";
 import { notificationRouter } from "./notification";
 import { adminRouter } from "./admin";
 
-export const appRouter = router({
+export const appRouter = createTRPCRouter({
   post: postRouter,
   profile: profileRouter,
   comment: commentRouter,
