+++ server/api/router/notification.ts
@@ -1,17 +1,17 @@
 import { TRPCError } from "@trpc/server";
-import { router, protectedProcedure } from "../trpc";
+import { createTRPCRouter, protectedProcedure } from "../trpc";
 import {
   GetNotificationsSchema,
   DeleteNotificationSchema,
 } from "../../../schema/notification";
 
-export const notificationRouter = router({
+export const notificationRouter = createTRPCRouter({
   delete: protectedProcedure
     .input(DeleteNotificationSchema)
     .mutation(async ({ input, ctx }) => {
       const { id } = input;
 
-      const currentNotification = await ctx.prisma.notification.findUnique({
+      const currentNotification = await ctx.db.notification.findUnique({
         where: { id },
       });
 
@@ -21,7 +21,7 @@ export const notificationRouter = router({
         });
       }
 
-      const notification = await ctx.prisma.notification.delete({
+      const notification = await ctx.db.notification.delete({
         where: {
           id,
         },
@@ -30,7 +30,7 @@ export const notificationRouter = router({
       return notification.id;
     }),
   deleteAll: protectedProcedure.mutation(async ({ ctx }) => {
-    const notification = await ctx.prisma.notification.deleteMany({
+    const notification = await ctx.db.notification.deleteMany({
       where: {
         userId: ctx.session.user.id,
       },
@@ -45,7 +45,7 @@ export const notificationRouter = router({
 
       const limit = input?.limit ?? 50;
       const { cursor } = input;
-      const response = await ctx.prisma.notification.findMany({
+      const response = await ctx.db.notification.findMany({
         take: limit + 1,
         where: {
           userId,
@@ -87,7 +87,7 @@ export const notificationRouter = router({
   getCount: protectedProcedure.query(async ({ ctx }) => {
     const userId = ctx?.session?.user?.id;
 
-    const count = await ctx.prisma.notification.count({
+    const count = await ctx.db.notification.count({
       where: {
         userId,
       },
