--- server/trpc/createRouter.ts
@@ -1,9 +1,0 @@
-import type { Context } from "./context";
-import * as trpc from "@trpc/server";
-
-/**
- * Helper function to create a router with context
- */
-export function createRouter() {
-  return trpc.router<Context>();
-}
