--- server/trpc/router/auth.ts
@@ -1,7 +1,0 @@
-import { router, publicProcedure } from "../trpc";
-
-export const authRouter = router({
-  getSession: publicProcedure.query(({ ctx }) => {
-    return ctx.session;
-  }),
-});
